// src/app/features/auth/data/dtos/user-response.dto.ts
export interface UserResponseDto {
  id: string;
  email: string;
  username: string;
  role_id: number;
  firstname: string;
  lastname: string;
  fullname: string;
  mobile_number: string;
  phone_number: string;
  line_id: string;
  location_id: string;
  status: number;
  is_superuser: boolean;
  verified: boolean;
  last_sign_in: string;
  createddate: string;
  updateddate: string;
}